requirements.txt
torch>=1.9.0+cpu
torchvision
transformers>=4.20.0
fastapi>=0.95.0
uvicorn>=0.22.0
streamlit>=1.24.0
pillow>=9.5.0
numpy>=1.24.0
sentence-transformers>=2.2.0

openai>=1.0.0
aiohttp>=3.8.4

pandas>=2.0.0
jsonschema>=4.0.0
lxml

pytesseract>=0.3.10
docx2txt>=0.8
python-pptx>=0.6.21
beautifulsoup4>=4.12.2

# Database Connectors
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
pymongo>=4.4.1
redis>=4.6.0

boto3>=1.28.0
google-cloud-storage>=2.9.0
azure-storage-blob>=12.17.0

safetensors>=0.3.1
einops>=0.6.1
tokenizers>=0.13.3
optimum>=1.8.5

pdfminer.six>=20221105
pymupdf>=1.22.3
markdown>=3.4.3

loguru>=0.7.0
tqdm>=4.65.0

# Security and Privacy
cryptography>=41.0.0
bcrypt>=4.1.0
regex>=2023.8.8

# Network and HTTP
requests>=2.31.0
urllib3>=2.0.0
watchdog>=3.0.0

pydantic
httpx
accelerate
qwen_vl_utils
bitsandbytes

# Web deployment
uvicorn
gunicorn
gradio

# Feishu integration
flask>=2.0.0
aiohttp>=3.8.0
requests>=2.25.0
opencv-python>=4.8.0

# Vector Database and Search
faiss-cpu>=1.7.0
chromadb>=0.4.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.2.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Jupyter and Tutorials
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Enterprise Integration
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4

# Async and Concurrency
asyncio-throttle>=1.0.2
aiofiles>=23.2.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# CLI and Terminal
rich>=13.6.0
click>=8.1.0
typer>=0.9.0

# File Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Optional: Advanced Features
fastapi>=0.95.0
openai>=1.3.0              # For OpenAI API integration
anthropic>=0.7.0           # For Anthropic API integration
langchain>=0.0.350         # For advanced LLM chains
langchain-community>=0.0.38

seaborn
scikit-learn